+++ src/main/java/open/commons/function/SQLConsumer.java
@@ -26,9 +26,14 @@
 
 package open.commons.function;
 
+import java.lang.reflect.Method;
 import java.sql.PreparedStatement;
 import java.sql.SQLException;
 import java.util.Objects;
+
+import open.commons.annotation.ColumnValue;
+import open.commons.utils.ObjectUtils;
+import open.commons.utils.SQLUtils;
 
 /**
  * Represents an operation that accepts a single input argument and returns no result. Unlike most other functional
@@ -99,5 +104,38 @@ public interface SQLConsumer<T> {
                 params[i] = null;
             }
         };
+    }
+
+    /**
+     * {@link PreparedStatement}에 데이터를 설정한다. <br>
+     * 해당 객체가 Primitive Type, Wrapper class Type 또는 {@link String} 인 경우 {@link #setParameters(Object...)} 로 전달되고,<br>
+     * 그외 Reference Type인 경우 {@link ColumnValue}가 적용된 {@link Method}를 이용하여 파라미터를 추출한다. <br>
+     * 
+     * <pre>
+     * [개정이력]
+     *      날짜    	| 작성자	|	내용
+     * ------------------------------------------
+     * 2020. 12. 22.		박준홍			최초 작성
+     * </pre>
+     *
+     * @param param
+     * @return
+     *
+     * @since 2020. 12. 22.
+     * @version 1.8.0
+     * @author Park_Jun_Hong_(fafanmama_at_naver_com)
+     * @see SQLUtils#setParameters(PreparedStatement, int, Object, String...)
+     */
+    public static SQLConsumer<PreparedStatement> setParameters(Object param, String... columnNames) {
+        if (((columnNames == null || columnNames.length < 1)) //
+                && (ObjectUtils.isPrimitive(param) //
+                        || ObjectUtils.isWrapper(param) //
+                        || String.class.equals(param.getClass()) //
+                ) //
+        ) {
+            return setParameters(new Object[] { param });
+        } else {
+            return stmt -> SQLUtils.setParameters(stmt, 0, param, columnNames);
+        }
     }
 }