+++ src/main/java/open/commons/utils/SQLUtils.java
@@ -45,16 +45,24 @@ import java.sql.Timestamp;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
+import java.util.Comparator;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.Map.Entry;
 import java.util.TreeMap;
+import java.util.function.BiFunction;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 
 import open.commons.TwoValueObject;
 import open.commons.annotation.ColumnDecl;
 import open.commons.annotation.ColumnDef;
+import open.commons.annotation.ColumnDef.ColumnNameType;
+import open.commons.annotation.ColumnValue;
 import open.commons.database.annotation.ColumnConstraint;
+import open.commons.function.TripleFunction;
 
 /**
  * 
@@ -62,6 +70,50 @@ import open.commons.database.annotation.ColumnConstraint;
  * @author Park_Jun_Hong_(fafanmama_at_naver_com)
  */
 public class SQLUtils {
+
+    public static final Pattern METHOD_BOOLEAN_PATTERN = Pattern.compile("^(is|get)(.+)$");
+    public static final Pattern METHOD_PATTERN = Pattern.compile("^(get)(.+)$");
+
+    /**
+     * 2개의 문자열을 대/소문자 비교여부에 따라서 비교
+     * 
+     * @param s1
+     *            비교할 문자열
+     * @param s2
+     *            비교할 문자열
+     * @param cs
+     *            대/소문자 비교 여부
+     * 
+     * @return 문자여 비교 결과
+     */
+    public static final TripleFunction<String, String, Boolean, Boolean> COLUMN_CHECKER = (s1, s2, cs) -> {
+        if (cs) {
+            return s1.equals(s2);
+        } else {
+            return s1.equalsIgnoreCase(s2);
+        }
+    };
+
+    /**
+     * {@link Method} 이름을 패턴 비교하여 컬럼명을 추출하여 제공한다.
+     * 
+     * @param ptn
+     *            {@link Method} 이름 비교 {@link Pattern}
+     * @param str
+     *            메소드 이름
+     * 
+     * @return 패턴과 매칭되지 않는 경우 <code>null</code>을 반환한다.
+     */
+    public static final BiFunction<Pattern, String, String> METHOD_MATCHER = (ptn, str) -> {
+        Matcher m = ptn.matcher(str);
+        if (m.matches()) {
+            return StringUtils.toLowerCase(m.group(2), 0);
+        } else {
+            return null;
+        }
+    };
+
+    public static final BiFunction<Class<?>, Class<ColumnValue>, List<Method>> COLUMN_VALUE_METHOD_PROVIDER = AnnotationUtils::getAnnotatedMethodsAll;
 
     /**
      * 
@@ -481,6 +533,112 @@ public class SQLUtils {
                 | SQLException e) {
             throw new SQLException("annotation: " + cdef + ", value: " + value, e);
         }
+    }
+
+    /**
+     * {@link PreparedStatement}에 주어진 객체에 포함된 데이터 중에서 전달받은(@param columnNames) 컬럼에 해당하는 값을 설정한다. <br>
+     * 
+     * <pre>
+     * [개정이력]
+     *      날짜    	| 작성자	|	내용
+     * ------------------------------------------
+     * 2020. 12. 22.		박준홍			최초 작성
+     * </pre>
+     *
+     * @param stmt
+     * @param index
+     * @param obj
+     * @param columnNames
+     * @return
+     * @throws SQLException
+     *
+     * @since 2020. 12. 22.
+     * @version 1.8.0
+     * @author Park_Jun_Hong_(fafanmama_at_naver_com)
+     */
+    public static int setParameters(PreparedStatement stmt, int index, Object obj, String... columnNames) throws SQLException {
+        // #1. @ColumnValue 어노테이션이 설정된 Method 조회
+        Class<?> type = obj.getClass();
+        List<Method> methods = SQLUtils.COLUMN_VALUE_METHOD_PROVIDER.apply(type, ColumnValue.class);
+
+        // #2. 사용자 지정 컬럼 여부에 따른 Method 필터링
+        if (columnNames == null || columnNames.length < 1) {
+            // ColumnValue#order() 값에 대한 오름차순 정렬 적용
+            methods.sort(Comparator.comparing(m -> {
+                return m.getAnnotation(ColumnValue.class).order();
+            }));
+        } else {
+            // 메소드별 컬럼명 대/소문자 비교 여부
+            final Map<String, Boolean> clmnCaseSensitive = new HashMap<>();
+            // 컬럼이름/메소드
+            final Map<String, Method> methodMap = CollectionUtils.toMapHSV(methods, m -> {
+                ColumnValue cv = m.getClass().getAnnotation(ColumnValue.class);
+                // 설정된 컬럼명이 빈 문자열이 경우 처리
+                String clmn = cv.name();
+                if (clmn.isEmpty()) {
+                    Class<?> rtnClass = m.getReturnType();
+
+                    if (boolean.class.isAssignableFrom(rtnClass) //
+                            || Boolean.class.isAssignableFrom(rtnClass)) {
+                        clmn = SQLUtils.METHOD_MATCHER.apply(SQLUtils.METHOD_BOOLEAN_PATTERN, m.getName());
+                    } else {
+                        clmn = SQLUtils.METHOD_MATCHER.apply(SQLUtils.METHOD_PATTERN, m.getName());
+                    }
+
+                    if (clmn == null) {
+                        throw new IllegalArgumentException(String.format("해당 데이터에 대한 컬럼명이 설정되지 않았습니다. 설정: %s, 메소드: %s", cv, m));
+                    }
+
+                    // begin - PATCH [2020. 9. 24.]: 컬럼명 타입에 따라 자동 변경 적용 |
+                    // Park_Jun_Hong_(fafanmama_at_naver_com)
+                    switch (cv.columnNameType()) {
+                        case CAMEL_CASE:
+                            clmn = StringUtils.toLowerCase(clmn, 0);
+                            break;
+                        case PASCAL_CASE:
+                            clmn = StringUtils.toPascalCase(clmn);
+                            break;
+                        case SNAKE_CASE:
+                            clmn = StringUtils.toSnakeCase(clmn);
+                            break;
+                        case NAME:
+                            // 그대로 사용
+                            break;
+                        default:
+                            throw new IllegalArgumentException(String.format("지원하지 않는 컬럼명 타입입니다. 지원: %s, 입력: %s", Arrays.toString(ColumnNameType.values()), cv.columnNameType()));
+                    }
+                    // end - Park_Jun_Hong_(fafanmama_at_naver_com), 2020. 9. 24.
+
+                }
+
+                clmnCaseSensitive.put(clmn, cv.caseSensitive());
+
+                return clmn;
+            }, m -> m);
+
+            methods = new ArrayList<>();
+
+            for (String clmn : columnNames) {
+                for (Entry<String, Method> entry : methodMap.entrySet()) {
+                    if (SQLUtils.COLUMN_CHECKER.apply(clmn, entry.getKey(), clmnCaseSensitive.get(entry.getKey()))) {
+                        methods.add(entry.getValue());
+                        break;
+                    }
+                }
+                throw new IllegalArgumentException(String.format("지원하지 않는 컬럼이름입니다. 컬럼명: %s, 메소드: %s", clmn, methodMap));
+            }
+        }
+
+        // #3. 필터링 된 Method를 이용하여 파라미터 값 설정
+        for (Method m : methods) {
+            try {
+                stmt.setObject(++index, m.invoke(obj));
+            } catch (IllegalAccessException | IllegalArgumentException | InvocationTargetException e) {
+                throw new SQLException("파라미터 설정 도중  에러가 발생하였습니다. 원인: " + e.getMessage(), e);
+            }
+        }
+
+        return index;
     }
 
     /**